Flx_is_totally_split: fix C++ signedness warning

+++ src/basemath/FpX_factor.c
@@ -717,7 +717,7 @@ Flx_is_totally_split(GEN f, ulong p)
   pari_sp av = avma;
   long n = degpol(f);
   if (n <= 1) return 1;
-  if (n > p) return 0;
+  if ((ulong)n > p) return 0;
   return gc_bool(av, Flx_is_totally_split_i(f,p));
 }
 int
